+++ satella/os/misc.py
@@ -1,5 +1,6 @@
 import os
 import sys
+import warnings
 
 
 def is_running_as_root() -> bool:
@@ -22,13 +23,14 @@ def suicide(kill_entire_pg: bool = True) -> None:
     """
     Kill self.
 
-    :param kill_entire_pg: whether to kill entire PG if a session leader
-    :raises NotImplementedError: called on Windows!
+    :param kill_entire_pg: whether to kill entire PG if a session leader. Won't work on Windows.
     """
 
     if sys.platform.startswith('win'):
-        raise NotImplementedError('suicide() not yet supported on Windows! (see issue #37)')
+        if kill_entire_pg:
+            warnings.warn('Windows does not support process leader groups', RuntimeWarning)
+        os.kill(os.getpid(), -9)
 
-    mypid = os.getpid()
-    kill = os.killpg if kill_entire_pg and os.getpgid(0) == mypid else os.kill
-    kill(mypid, 9)
+    my_pid = os.getpid()
+    kill = os.killpg if kill_entire_pg and os.getpgid(0) == my_pid else os.kill
+    kill(my_pid, 9)